intermediate refactor of package names
