+++ src/databricks/labs/lakebridge/config.py
@@ -8,8 +8,8 @@ from typing import Any, cast
 
 from databricks.labs.blueprint.installation import JsonValue
 from databricks.labs.blueprint.tui import Prompts
-from databricks.labs.remorph.transpiler.transpile_status import TranspileError
-from databricks.labs.remorph.reconcile.recon_config import Table
+from databricks.labs.lakebridge.transpiler.transpile_status import TranspileError
+from databricks.labs.lakebridge.reconcile.recon_config import Table
 
 
 logger = logging.getLogger(__name__)
