+++ src/databricks/labs/lakebridge/deployment/installation.py
@@ -11,7 +11,7 @@ from databricks.sdk.errors import NotFound
 from databricks.sdk.mixins.compute import SemVer
 from databricks.sdk.errors.platform import InvalidParameterValue, ResourceDoesNotExist
 
-from databricks.labs.remorph.config import RemorphConfigs
+from databricks.labs.lakebridge.config import RemorphConfigs
 from databricks.labs.remorph.deployment.recon import ReconDeployment
 
 logger = logging.getLogger("databricks.labs.remorph.install")
