+++ src/databricks/labs/lakebridge/helpers/recon_config_utils.py
@@ -1,7 +1,7 @@
 import logging
 
 from databricks.labs.blueprint.tui import Prompts
-from databricks.labs.remorph.reconcile.constants import ReconSourceType
+from databricks.labs.lakebridge.reconcile.constants import ReconSourceType
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.errors.platform import ResourceDoesNotExist
