+++ src/databricks/labs/lakebridge/reconcile/recon_capture.py
@@ -8,7 +8,7 @@ from pyspark.sql.types import StringType, StructField, StructType
 from pyspark.errors import PySparkException
 from sqlglot import Dialect
 
-from databricks.labs.remorph.config import DatabaseConfig, Table, ReconcileMetadataConfig
+from databricks.labs.lakebridge.config import DatabaseConfig, Table, ReconcileMetadataConfig
 from databricks.labs.remorph.reconcile.recon_config import TableThresholds
 from databricks.labs.remorph.transpiler.sqlglot.dialect_utils import get_key_from_dialect
 from databricks.labs.remorph.reconcile.exception import (
