+++ src/databricks/labs/lakebridge/uninstall.py
@@ -3,10 +3,10 @@ import logging
 from databricks.labs.blueprint.entrypoint import is_in_debug
 from databricks.sdk import WorkspaceClient
 
-from databricks.labs.remorph.__about__ import __version__
-from databricks.labs.remorph.contexts.application import ApplicationContext
+from databricks.labs.lakebridge.__about__ import __version__
+from databricks.labs.lakebridge.contexts.application import ApplicationContext
 
-logger = logging.getLogger("databricks.labs.remorph.install")
+logger = logging.getLogger("databricks.labs.lakebridge.install")
 
 
 def run(context: ApplicationContext):
