+++ src/databricks/labs/lakebridge/upgrades/v0.6.0_alter_metrics_datatype.py
@@ -4,14 +4,14 @@ import logging
 from databricks.labs.blueprint.installation import Installation
 from databricks.sdk import WorkspaceClient
 
-from databricks.labs.remorph.contexts.application import ApplicationContext
-from databricks.labs.remorph.deployment.upgrade_common import (
+from databricks.labs.lakebridge.contexts.application import ApplicationContext
+from databricks.labs.lakebridge.deployment.upgrade_common import (
     current_table_columns,
     installed_table_columns,
     check_table_mismatch,
     recreate_table_sql,
 )
-from databricks.labs.remorph.helpers import db_sql
+from databricks.labs.lakebridge.helpers import db_sql
 
 logger = logging.getLogger(__name__)
